# docs/swagger.yaml
basePath: /api/v1
definitions:
  handler.articleResp:
    properties:
      author_id:
        type: string
      body:
        type: string
      created_at:
        type: string
      id:
        type: string
      state:
        type: string
      title:
        type: string
      updated_at:
        type: string
    type: object
  handler.createArticleReq:
    properties:
      body:
        type: string
      title:
        maxLength: 500
        minLength: 1
        type: string
    required:
    - body
    - title
    type: object
  handler.loginReq:
    properties:
      email:
        type: string
      password:
        minLength: 8
        type: string
    required:
    - email
    - password
    type: object
  handler.refreshReq:
    properties:
      refresh_token:
        type: string
    required:
    - refresh_token
    type: object
  handler.registerReq:
    properties:
      email:
        type: string
      password:
        minLength: 8
        type: string
    required:
    - email
    - password
    type: object
  handler.tokenResp:
    properties:
      access_token:
        type: string
      expires_at:
        type: string
      refresh_token:
        type: string
    type: object
  handler.updateArticleReq:
    properties:
      body:
        type: string
      state:
        enum:
        - draft
        - published
        type: string
      title:
        maxLength: 500
        minLength: 1
        type: string
    required:
    - body
    - state
    - title
    type: object
host: localhost:8080
info:
  contact: {}
  description: REST API with JWT, RBAC, and articles CRUD.
  title: Gox Boilerplate API
  version: "1.0"
paths:
  /articles:
    get:
      parameters:
      - default: 20
        description: Limit
        in: query
        name: limit
        type: integer
      - default: 0
        description: Offset
        in: query
        name: offset
        type: integer
      - description: Only my articles
        in: query
        name: mine
        type: boolean
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/handler.articleResp'
            type: array
      security:
      - BearerAuth: []
      summary: List articles
      tags:
      - articles
    post:
      consumes:
      - application/json
      parameters:
      - description: payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handler.createArticleReq'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/handler.articleResp'
      security:
      - BearerAuth: []
      summary: Create article (draft)
      tags:
      - articles
  /articles/{id}:
    delete:
      parameters:
      - description: Article ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "204":
          description: No Content
      security:
      - BearerAuth: []
      summary: Delete article (admin)
      tags:
      - articles
    get:
      parameters:
      - description: Article ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.articleResp'
      security:
      - BearerAuth: []
      summary: Get article by ID
      tags:
      - articles
    put:
      consumes:
      - application/json
      parameters:
      - description: Article ID
        in: path
        name: id
        required: true
        type: string
      - description: payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handler.updateArticleReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.articleResp'
      security:
      - BearerAuth: []
      summary: Update article
      tags:
      - articles
  /articles/{id}/publish:
    post:
      parameters:
      - description: Article ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.articleResp'
      security:
      - BearerAuth: []
      summary: Publish article (admin)
      tags:
      - articles
  /auth/login:
    post:
      consumes:
      - application/json
      parameters:
      - description: credentials
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handler.loginReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.tokenResp'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Login
      tags:
      - auth
  /auth/logout:
    post:
      responses:
        "204":
          description: No Content
      security:
      - BearerAuth: []
      summary: Logout (blacklist access token)
      tags:
      - auth
  /auth/refresh:
    post:
      consumes:
      - application/json
      parameters:
      - description: payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handler.refreshReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.tokenResp'
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Refresh tokens
      tags:
      - auth
  /auth/register:
    post:
      consumes:
      - application/json
      parameters:
      - description: payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handler.registerReq'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/handler.tokenResp'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Register
      tags:
      - auth
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"